requirements.txt
# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.25.2

# Google Generative AI (Gemini)
google-generativeai>=0.5.0

# Audio processing
# Use latest pydub which handles Python 3.13 better
pydub>=0.25.1
# Alternative audio processing that doesn't require audioop
ffmpeg-python>=0.2.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Progress bar
tqdm==4.66.1

# Additional dependencies
httplib2==0.22.0
cachetools==5.3.2
protobuf==4.25.1

# OAuth2 dependencies
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Google API core dependencies
googleapis-common-protos==1.62.0
google-api-core==2.15.0